@@ -1,6 +1,6 @@
-module github.com/talos-systems/go-blockdevice
+module github.com/siderolabs/go-blockdevice
 
-go 1.14
+go 1.19
 
 require (
 	github.com/google/uuid v1.1.2
@@ -10,4 +10,11 @@ require (
 	github.com/talos-systems/go-retry v0.1.1-0.20201113203059-8c63d290a688
 	golang.org/x/sys v0.0.0-20201130171929-760e229fe7c5
 	golang.org/x/text v0.3.4
+)
+
+require (
+	github.com/armon/circbuf v0.0.0-20190214190532-5111143e8da2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
 )
