@@ -2,14 +2,14 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-//nolint: scopelint
+//nolint:scopelint
 package endianness_test
 
 import (
 	"reflect"
 	"testing"
 
-	"github.com/talos-systems/go-blockdevice/blockdevice/endianness"
+	"github.com/siderolabs/go-blockdevice/blockdevice/endianness"
 )
 
 var (
